[CBRD-25542] Problem with partition pruning when using multi-column indexes with IN and range predicates (#5689)
http://jira.cubrid.org/browse/CBRD-25542

Partition pruning does not work correctly when the WHERE clause contains an IN condition on partition key column A,
combined with a range condition (> or <) on column B without an equality condition, where both columns are part of the same index. 

This specifically occurs when the partitioned table has an index composed of columns A, B, and other columns, and this index is used to scan the partitioned table.

The issue is caused by the use of a multi-column index.

With the addition of the MVCC feature, the where_range now contains a PRED_EXPR that compares a single column of the index. This has been modified to utilize where_range for partition pruning.